+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - ... <!-- Add new lines here. Version number will be decided later -->
 - **BREAKING** feat: Upgrade `rollup` to v2
 - **BREAKING** feat: Switch to `@rollup/plugin-typescript`
+- feat: Auto-tidy TypeScript declarations – unless
+  `typescriptOpts.rawDeclarations` is `true`
 
 ## 0.3.23 – 0.3.25